Add protobuf workaround for upstream breaking change
The latest version of the protocol buffers library has a breaking change which is
present in the current versions of the Pulumi toolchain. This adds a temporary
workaround while those version mismatches are ironed out so that we can continue to run
our projects without having to pepper this environment setting across our pipelines.