chore: Add build-time install of Pulumi plugins
During our Pipeline runs we are installing the plugins twice every time the pipeline
executes. This leads to a lot of Github rate limits being exceeded and slower build
times for no good reason. This adds a one-liner to install the Pulumi plugins that we
depend on based on the Python dependencies.